Refactored matching engine tests to use book and queue as dependency

+++ Application/Common/Models/StopOrderProcessingResult.cs
@@ -2,7 +2,7 @@ namespace Application.Common.Models;
 
 public sealed class StopOrderProcessingResult
 {
-    public required int FilledCount { get; init; } 
+    public required decimal FilledCount { get; init; } 
     public required int MatchedCount { get; init; } 
     public required decimal CostAccumulate { get; init; } 
 }
